k8s: RBAC improvements
This patch contains several RBAC improvements that are intermixed.

First, seperate RBAC installation into a different file so it can be used inside
an ansible loop more easily.

The RBAC setup now reads the uid/gid of the tool account and injects it into the PSP.
Also, an additional rolebinding is introduced for the namespace's default service account.

While at it, set the unix user primary group.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>