toolforge: install_tool_unix_account: generalice algorithm
Specify bash as shell, so we get consistency across systems.
Also, don't use `getent`, which is only available on a subset of Unix systems.
And, use integers as id fallback values, so the `test` function works as
expected.

In hopes that this is equivalent but also works on mac OSX.

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>