lima-kilo: introduce basic CLI support
Introduce a role capable of deploying a lima-kilo flavor of the toolforge-jobs
CLI.

This is very similar to the testing setup that jobs-framework-cli uses, so we
can drop the code from there as next step.

With this patch, the user can do:

$ ~/.local/toolforge-lima-kilo/bin/toolforge-jobs list
WARNING: not running as the tool account? Likely to fail. Perhaps you forgot `become <tool>`?
[..]

(if jobs-api is deployed in the k8s cluster)

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>