k8s: add stronger toolforge-deploy support
This deploys the selected components as they are deployed via toolforge-deploy.

The side benefit is that we can leverage harbor charts/images and no longer 'build'
the container images locally, given the container images can be reused from the CI.

If a developer wants to override this, because is developing a given custom component,
it should play with a local git checkout of the given component repo, and do the
build manually outside lima-kilo (exactly as before).

Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>