start-devenv.sh: don't hardcode the VM name
this allows the VM name to be specified as an option
(will default to "lima-kilo" if not specified)

it only changes the name of the lima-vm instance,
internal hostname and other identifiers remain
as `lima-kilo`

this is useful if you want to have multiple lima-kilo
instances, e.g. one more stable for testing MRs,
others for testing changes to lima-kilo itself, etc.