toolforge_deploy_mr: use always toolsbeta as repo for mr deploys
MRs don't get pushed to tools, and sometimes we want to deploy the
tools version of something by default (so the repo in local.yaml* is
tools). This makes sure that if you deploy an MR, the repo is going to
be toolsbeta too, no matter what's the original repo config.

Signed-off-by: David Caro <dcaro@wikimedia.org>