install-binary-from-url - cache binaries in cache drive
Currently on every ansible run the binaries are re-downloaded into a
temporary directory and then discarded.

When operating on a lower bandwidth connection this consumes a lot of
resources, with partial failures causing re-downloading and very long
convergence times.

Download instead to the cache drive so they can be reused from there
instead of redownloaded.

Bug: T402684
Signed-off-by: David Caro <dcaro@wikimedia.org>