install-binary-from-url - add checksums for dest
Currently on every ansible run the binaries are re-downloaded into a
tempoary directory and then discarded.

When operating on a lower bandwidth connection this consumes a lot of
resources, with partial failures causing re-downloading and very long
convergence times.

Introduce a checksum for the target binary (potentially the same as the
download checksum, or the extracted file checksum) and gate the creation
of temp dir, download, extract, copy on this matching the existing file.