deploy components - don't report as changed
Currently toolforge-deploy/deploy.sh only applies when there are changes
(between the repo and running), however this is not exposed in a way
that easily allows ansible to check for a diff then execute on changes,
which would correctly report changes for the run.

Report no changes from the task so the ansible run output is clean
during a subsequent execution.