Configure nsswitch.conf after sssd is running
Currently we configure nsswitch to use sssd prior to the sssd service
being installed.

On installation of the sssd service, the nsswitch gets modified.

Once kind is setup, the file is bind-mounted into a docker container
(extraMounts).

If ansible is run a second time, the nsswitch config is copied,
as the contents differs.

Unfortunatly for processes inside the bind-mind (kubelet etc) the file
is now not accessible as the old file (inode) has been removed.

This causes jobs-api to fail to restart with a 'no such file or
directory' error, which isn't obuvious to debug as within the container
the file is accessible and updated.

Changing the ordering here avoids the contents being changed in a
simplistic manner. There are other cases where this might get changed
(which restarting docker will fix), but at this point in the ansible
flow docker is not installed, so for now let's go simple rather than
introducing a complex handler.