deploy_mr - Use project/job as returned by the API
Bug: T407983

When a non-maintainer contributes a change it exists on a fork under a
different project than that returned by `get_project` (upstream).

1.
During `get_chart_job` use the `project_id` contained within the
`pipeline` response data, which can be either the same as `project.id`
(maintainer change) or different (non-maintainer change).

This allows the pipeline jobs to be successfully returned, rather than
returning a 404, or worse, an un-related change logs.

2.
During the log fetching, use the base URL returned from `get_chart_job`,
which has the correct project slug and job id encoded, rather than
constructing it with a constant value.

This allows non-maintainer jobs to be found, rather than a 404